ricalcoli affluenze funzionano aggiunte librerie jquery table

+++ src/main/java/com/deltasi/elezioni/contracts/ISezioneService.java
@@ -9,6 +9,6 @@ public interface ISezioneService {
     Sezione findById(Integer id);
     Sezione findByNumerosezioneAndTipoelezioneId(Integer numeroSezione, Integer tipoElezioneId);
     List<Sezione> findAllBy();
-    List<Integer> countAllByTipoelezioneIdAndTipoelezioneIdIn(int tipoElezione, int tipoElezione1);
+    List<Long> countAllByTipoelezioneIdAndTipoelezioneIdIn(int tipoElezione, int tipoElezione1);
     List<Long> countAllByTipoelezioneIdAndMunicipioAndTipoelezioneIdIn(int tipoElezioneId, int municipio,int tipoElezione1);
 }
